weather-data/requirements.txt
Flask==3.0.3
requests==2.32.3
cassandra-driver==3.29.1